Split READMEs & helper modules.
