+++ README.md
@@ -21,27 +21,5 @@ pre-commit install
 ```
 
 ## Experimental Scripts
-Experimenatal scripts are located within the `scripts` directory.
-
-### Setup and installation
-Scripts may require their own specific environment.
-These are specified as conda `.yml` files and are located within `scripts/envs`.
-
-For example, to create and activate the `ppa-ocr` environment, run the following commands:
-```
-conda env create -f scripts/envs/ppa-ocr.yml
-conda activate ppa-ocr
-```
-
-### Script descriptions
-
-#### get_character_stats.py
-This script compiles character-level statistics for a PPA text (sub)corpus.
-It requires the `ppa-ocr` environment.
-This script has an additional dependency, it requires the [en_core_web_lg](https://spacy.io/models/en#en_core_web_lg) spacy language model, which can be downloaded by running the following command:
-```
-python -m spacy download en_core_web_lg
-````
-#### evaluate_ocr.py
-This script compiles OCR quality statistics for a PPA text (sub)corpus
-It requiers the `ppa-ocr` environment.
+Experimental scripts associated with `corppa` are located within the `scripts` directory.
+See this directory's README for more detail.
